+++ test/unit/specs/components/governance/ModalPropose.spec.js
@@ -48,6 +48,19 @@ describe(`ModalPropose`, () => {
     expect(wrapper.vm.$refs.actionModal.open).toHaveBeenCalled()
   })
 
+  it(`clears on close`, () => {
+    wrapper.setData({ title: `test`, description: `test`, amount: 100000000 })
+    // produce validation error as amount is too high
+    wrapper.vm.$v.$touch()
+    expect(wrapper.vm.$v.$error).toBe(true)
+
+    wrapper.vm.clear()
+    expect(wrapper.vm.$v.$error).toBe(false)
+    expect(wrapper.vm.title).toBe(``)
+    expect(wrapper.vm.description).toBe(``)
+    expect(wrapper.vm.amount).toBe(0)
+  })
+
   describe(`default values are set correctly`, () => {
     it(`the proposal type defaults to 'Text'`, () => {
       expect(wrapper.vm.type).toEqual(`Text`)