+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - [\#2043](https://github.com/cosmos/voyager/issues/2043) removed no-balance modals on staking page @fedekunze
 - [\#2161](https://github.com/cosmos/voyager/pull/2161) fixed status dot position @jbibla
+- [\#2193](https://github.com/cosmos/voyager/issues/2193) hide seed import @faboweb
 
 ## [1.0.0-beta.14] - 2019-03-06